requirements.txt
marshmallow==3.21.1
git+https://github.com/YaserFazil/dynamorm.git
boto3==1.34.81
requests==2.31.0
python-dotenv==1.0.1
Flask==3.0.3
Flask-Cors==4.0.1
gunicorn==22.0.0